requirements.txt
face_recognition==1.3.5
opencv-python==4.8.1.78
mysql-connector-python==8.2.0
numpy==1.24.3
Pillow==10.0.0
